Eliminate GHC.Internal.Base_LHAssumptions
